more start up node id and port checks

+++ abrim/node.py
@@ -14,33 +14,35 @@ def output_reader(proc, prefix):
 
 def main():
     node_id, client_port = args_init()
+    if not node_id or not client_port:
+        pass
+    else:
+        proc_ui, thread_ui = launch_subprocess('ui.py', "UI___", node_id, client_port)
+        proc_queue_in, thread_queue_in = launch_subprocess('input.py', "INPUT", node_id, client_port + 1)
+        proc_queue_out, thread_queue_out = launch_subprocess('out.py', "OUT__", node_id, client_port + 1)
+        proc_queue_patch, thread_queue_patch = launch_subprocess('patch.py', "PATCH", node_id, client_port + 1)
 
-    proc_ui, thread_ui = launch_subprocess('ui.py', "UI___", node_id, client_port)
-    proc_queue_in, thread_queue_in = launch_subprocess('input.py', "INPUT", node_id, client_port + 1)
-    proc_queue_out, thread_queue_out = launch_subprocess('out.py', "OUT__", node_id, client_port + 1)
-    proc_queue_patch, thread_queue_patch = launch_subprocess('patch.py', "PATCH", node_id, client_port + 1)
+        webbrowser.open_new_tab("http://localhost:" + str(client_port) + "/")
 
-    webbrowser.open_new_tab("http://localhost:" + str(client_port) + "/")
+        try:
+            while True:
+                time.sleep(0.1)
+            #
+            # for i in range(600):
+            #     #print("test")
+            #     time.sleep(0.1)
+        finally:
+            # This is in 'finally' so that we can terminate the child if something
+            # goes wrong
+            proc_terminate(proc_ui)
+            proc_terminate(proc_queue_in)
+            proc_terminate(proc_queue_out)
+            proc_terminate(proc_queue_patch)
 
-    try:
-        while True:
-            time.sleep(0.1)
-        #
-        # for i in range(600):
-        #     #print("test")
-        #     time.sleep(0.1)
-    finally:
-        # This is in 'finally' so that we can terminate the child if something
-        # goes wrong
-        proc_terminate(proc_ui)
-        proc_terminate(proc_queue_in)
-        proc_terminate(proc_queue_out)
-        proc_terminate(proc_queue_patch)
-
-    thread_ui.join()
-    thread_queue_in.join()
-    thread_queue_out.join()
-    thread_queue_patch.join()
+        thread_ui.join()
+        thread_queue_in.join()
+        thread_queue_out.join()
+        thread_queue_patch.join()
 
 
 def proc_terminate(my_proc):
